+++ src/core/sort/newsort.cc
@@ -26,7 +26,7 @@
 #include "python/args.h"       // py::PKArgs
 #include "sort/common.h"       // array
 #include "sort/insert-sort.h"  // insert_sort
-#include "sort/radixsort.h"    // RadixSort
+#include "sort/radix-sort.h"   // RadixSort
 #include "sort/sorter_bool.h"  // Sorter_Bool
 #include "utils/assert.h"      // xassert
 #include "utils/misc.h"        // dt::nlz
